start leaflet cleanup, rest will comein 0.16
