+++ src/code/leafletDrawImports.js
@@ -1,6 +1,6 @@
 // Put anything from leaflet-draw that we need in here.
 
-const Tooltip = L.Class.extend({
+export const Tooltip = L.Class.extend({
   initialize: function(map) {
     this._map = map;
     this._popupPane = map._panes.popupPane;
@@ -62,85 +62,30 @@ const Tooltip = L.Class.extend({
 });
 
 export const Feature = L.Handler.extend({
+  // includes: L.Evented,
   includes: L.Mixin.Events,
 
   initialize: function(map, options) {
     this._map = map;
     this._container = map._container;
-    this._overlayPane = map._panes.overlayPane;
-    this._popupPane = map._panes.popupPane;
-
-    // Merge default shapeOptions options with custom shapeOptions
-    if (options && options.shapeOptions) {
-      options.shapeOptions = L.Util.extend(
-        {},
-        this.options.shapeOptions,
-        options.shapeOptions
-      );
-    }
     L.Util.extend(this.options, options);
   },
 
   enable: function() {
-    if (this._enabled) {
-      return;
-    }
-
+    if (this._enabled) return;
     L.Handler.prototype.enable.call(this);
     this.fire("enabled", { handler: this.type });
-    this._map.fire("draw:drawstart", { layerType: this.type });
   },
 
   disable: function() {
-    if (!this._enabled) {
-      return;
-    }
-
+    if (!this._enabled) return;
     L.Handler.prototype.disable.call(this);
     this.fire("disabled", { handler: this.type });
-    this._map.fire("draw:drawstop", { layerType: this.type });
   },
 
-  addHooks: function() {
-    if (this._map) {
-      L.DomUtil.disableTextSelection();
+  addHooks: function() {},
 
-      this._tooltip = new Tooltip(this._map);
-
-      L.DomEvent.addListener(
-        this._container,
-        "keyup",
-        this._cancelDrawing,
-        this
-      );
-    }
-  },
-
-  removeHooks: function() {
-    if (this._map) {
-      L.DomUtil.enableTextSelection();
-
-      this._tooltip.dispose();
-      this._tooltip = null;
-
-      L.DomEvent.removeListener(this._container, "keyup", this._cancelDrawing);
-    }
-  },
-
-  setOptions: function(options) {
-    L.setOptions(this, options);
-  },
-
-  _fireCreatedEvent: function(layer) {
-    this._map.fire("draw:created", { layer: layer, layerType: this.type });
-  },
-
-  // Cancel drawing when the escape key is pressed
-  _cancelDrawing: function(e) {
-    if (e.keyCode === 27) {
-      this.disable();
-    }
-  }
+  removeHooks: function() {}
 });
 
 export const Toolbar = L.Class.extend({
